Move NTP message code to Rex::Proto::NTP, simplify option handling
