+++ lib/rex/proto/ntp.rb
@@ -1,0 +1,3 @@
+# -*- coding: binary -*-
+require 'rex/proto/ntp/constants'
+require 'rex/proto/ntp/modes'
